fixes based on code review, found some bugs in tests that I fixed
